pyproject.toml
[project]
name = "translator"
version = "0.1.0"
description = "Translate PDFs to Farsi and save as PDFs"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Your Team" }]

dependencies = [
  "openai>=1.40.0",
  "pymupdf>=1.24.0",
  "arabic-reshaper>=3.0.0",
  "python-bidi>=0.4.2",
]

[project.scripts]
translate-pdf = "translator.cli:main"
